# Faster sparse vectors plain search