This PR optimizes the plain search through the inverted index for sparse vectors.

```
sparse-vector-search-group/inverted-index-filtered-plain
                        time:   [134.17 ms 134.44 ms 134.70 ms]
                        change: [-34.599% -34.462% -34.317%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
...
sparse-vector-search-group/plain-filtered-payload-index
                        time:   [128.26 ms 128.41 ms 128.56 ms]
                        change: [-35.495% -35.337% -35.185%] (p = 0.00 < 0.05)
                        Performance has improved.


```

The insight is that by sorting the input ids, we can actually fully leverage the posting list iterator by performing binary search on a shrinking space.