@agourlay 2023-12-14
It seems the Python congruence tests do not like this change.
I need to investigate this before merging.